+++ library/src/schemas/enum/enum.ts
@@ -2,9 +2,9 @@ import type { BaseSchema, ErrorMessage } from '../../types.ts';
 import { getSchemaIssues, getOutput } from '../../utils/index.ts';
 
 /**
- * Native enum type.
+ * Enum type.
  */
-export type NativeEnum = {
+export type Enum = {
   [key: string]: string | number;
   [key: number]: string;
 };
@@ -12,36 +12,36 @@ export type NativeEnum = {
 /**
  * Native enum schema type.
  */
-export type NativeEnumSchema<
-  TNativeEnum extends NativeEnum,
-  TOutput = TNativeEnum[keyof TNativeEnum]
-> = BaseSchema<TNativeEnum[keyof TNativeEnum], TOutput> & {
-  type: 'native_enum';
-  nativeEnum: TNativeEnum;
+export type EnumSchema<
+  TEnum extends Enum,
+  TOutput = TEnum[keyof TEnum]
+> = BaseSchema<TEnum[keyof TEnum], TOutput> & {
+  type: 'enum';
+  enum: TEnum;
 };
 
 /**
  * Creates a enum schema.
  *
- * @param nativeEnum The native enum value.
+ * @param enum_ The enum value.
  * @param error The error message.
  *
  * @returns A enum schema.
  */
-export function nativeEnum<TNativeEnum extends NativeEnum>(
-  nativeEnum: TNativeEnum,
+export function enum_<TEnum extends Enum>(
+  enum_: TEnum,
   error?: ErrorMessage
-): NativeEnumSchema<TNativeEnum> {
+): EnumSchema<TEnum> {
   return {
     /**
      * The schema type.
      */
-    type: 'native_enum',
+    type: 'enum',
 
     /**
-     * The native enum value.
+     * The enum value.
      */
-    nativeEnum,
+    enum: enum_,
 
     /**
      * Whether it's async.
@@ -58,18 +58,25 @@ export function nativeEnum<TNativeEnum extends NativeEnum>(
      */
     _parse(input, info) {
       // Check type of input
-      if (!Object.values(nativeEnum).includes(input as any)) {
+      if (!Object.values(enum_).includes(input as any)) {
         return getSchemaIssues(
           info,
           'type',
-          'native_enum',
+          'enum',
           error || 'Invalid type',
           input
         );
       }
 
       // Return input as output
-      return getOutput(input as TNativeEnum[keyof TNativeEnum]);
+      return getOutput(input as TEnum[keyof TEnum]);
     },
   };
 }
+
+/**
+ * See {@link enum_}
+ *
+ * @deprecated Use `enum_` instead.
+ */
+export const nativeEnum = enum_;