+++ library/src/schemas/enum/enumAsync.test.ts
@@ -1,15 +1,15 @@
 import { describe, expect, test } from 'vitest';
 import { parseAsync } from '../../methods/index.ts';
-import { nativeEnumAsync } from './nativeEnumAsync.ts';
+import { enumAsync } from './enumAsync.ts';
 
 enum Direction {
   Left,
   Right,
 }
 
-describe('nativeEnumAsync', () => {
+describe('enumAsync', () => {
   test('should pass only enum values', async () => {
-    const schema = nativeEnumAsync(Direction);
+    const schema = enumAsync(Direction);
     const input1 = Direction.Left;
     const output1 = await parseAsync(schema, input1);
     expect(output1).toBe(input1);
@@ -24,7 +24,7 @@ describe('nativeEnumAsync', () => {
   test('should throw custom error', async () => {
     const error = 'Value is not a enum value!';
     await expect(
-      parseAsync(nativeEnumAsync(Direction, error), 'test')
+      parseAsync(enumAsync(Direction, error), 'test')
     ).rejects.toThrowError(error);
   });
 });