+++ library/src/schemas/enum/enumAsync.ts
@@ -1,40 +1,40 @@
 import type { BaseSchemaAsync, ErrorMessage } from '../../types.ts';
 import { getSchemaIssues, getOutput } from '../../utils/index.ts';
-import type { NativeEnum } from './nativeEnum.ts';
+import type { Enum } from './enum.ts';
 
 /**
  * Native enum schema async type.
  */
-export type NativeEnumSchemaAsync<
-  TNativeEnum extends NativeEnum,
-  TOutput = TNativeEnum[keyof TNativeEnum]
-> = BaseSchemaAsync<TNativeEnum[keyof TNativeEnum], TOutput> & {
-  type: 'native_enum';
-  nativeEnum: TNativeEnum;
+export type EnumSchemaAsync<
+  TEnum extends Enum,
+  TOutput = TEnum[keyof TEnum]
+> = BaseSchemaAsync<TEnum[keyof TEnum], TOutput> & {
+  type: 'enum';
+  enum: TEnum;
 };
 
 /**
  * Creates an async enum schema.
  *
- * @param nativeEnum The native enum value.
+ * @param enum_ The enum value.
  * @param error The error message.
  *
  * @returns An async enum schema.
  */
-export function nativeEnumAsync<TNativeEnum extends NativeEnum>(
-  nativeEnum: TNativeEnum,
+export function enumAsync<TEnum extends Enum>(
+  enum_: TEnum,
   error?: ErrorMessage
-): NativeEnumSchemaAsync<TNativeEnum> {
+): EnumSchemaAsync<TEnum> {
   return {
     /**
      * The schema type.
      */
-    type: 'native_enum',
+    type: 'enum',
 
     /**
-     * The native enum value.
+     * The enum value.
      */
-    nativeEnum,
+    enum: enum_,
 
     /**
      * Whether it's async.
@@ -51,18 +51,25 @@ export function nativeEnumAsync<TNativeEnum extends NativeEnum>(
      */
     async _parse(input, info) {
       // Check type of input
-      if (!Object.values(nativeEnum).includes(input as any)) {
+      if (!Object.values(enum_).includes(input as any)) {
         return getSchemaIssues(
           info,
           'type',
-          'native_enum',
+          'enum',
           error || 'Invalid type',
           input
         );
       }
 
       // Return input as output
-      return getOutput(input as TNativeEnum[keyof TNativeEnum]);
+      return getOutput(input as TEnum[keyof TEnum]);
     },
   };
 }
+
+/**
+ * See {@link enumAsync}
+ *
+ * @deprecated Use `enumAsync` instead.
+ */
+export const nativeEnumAsync = enumAsync;